[FP-6984] Add cron job to execute enqueued tasks (#7961)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical details and possible side effects. -->
I moved the async `/private/protected/task/execute_tasks` to a synchronous cron job.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->
Tested locally -- see comments.

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->

+++ backend/components/api/core/src/task/mod.rs
@@ -40,7 +40,7 @@ pub fn execute_webhook_tasks(state: State) {
     poll_and_execute_tasks_non_blocking(state, 10, Some(TaskKind::FireWebhook))
 }
 
-pub fn poll_and_execute_tasks_non_blocking(state: State, limit: i64, kind: Option<TaskKind>) {
+pub fn poll_and_execute_tasks_non_blocking(state: State, limit: u32, kind: Option<TaskKind>) {
     let fut = async move {
         let _ = poll_and_execute_tasks(&state, limit, kind)
             .await
@@ -60,7 +60,7 @@ pub fn poll_and_execute_tasks_non_blocking(state: State, limit: i64, kind: Optio
 
 pub async fn poll_and_execute_tasks(
     state: &State,
-    limit: i64,
+    limit: u32,
     kind: Option<TaskKind>,
 ) -> Result<Vec<Task>, DbError> {
     let tasks: Vec<(Task, TaskExecution)> = state
